Reorganized files and installed react-swipeable and framer-motion
