+++ package.json
@@ -12,9 +12,11 @@
   "dependencies": {
     "@vis.gl/react-google-maps": "^1.0.0",
     "classnames": "^2.5.1",
+    "framer-motion": "^11.2.6",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-icons": "^5.2.1"
+    "react-icons": "^5.2.1",
+    "react-swipeable": "^7.0.1"
   },
   "devDependencies": {
     "@types/react": "^18.2.66",
